+++ Admin/AbstractAdmin.php
@@ -416,14 +416,7 @@ abstract class AbstractAdmin implements AdminInterface, DomainObjectInterface
 
     protected $cacheIsGranted = array();
 
-    protected $listModes = array(
-        'list' => array(
-            'class' => 'fa fa-list fa-fw',
-        ),
-        'mosaic' => array(
-            'class' => 'fa fa-th-large fa-fw',
-        ),
-    );
+    protected $listModes;
 
     /**
      * The Access mapping.
@@ -530,6 +523,8 @@ abstract class AbstractAdmin implements AdminInterface, DomainObjectInterface
 
         $this->predefinePerPageOptions();
         $this->datagridValues['_per_page'] = $this->maxPerPage;
+
+        $this->initListModes();
     }
 
     /**
@@ -1055,10 +1050,7 @@ abstract class AbstractAdmin implements AdminInterface, DomainObjectInterface
         }
 
         foreach ($this->getExtensions() as $extension) {
-            // TODO: remove method check in next major release
-            if (method_exists($extension, 'configureBatchActions')) {
-                $actions = $extension->configureBatchActions($this, $actions);
-            }
+            $actions = $extension->configureBatchActions($this, $actions);
         }
 
         return $actions;
@@ -2802,10 +2794,7 @@ abstract class AbstractAdmin implements AdminInterface, DomainObjectInterface
         $list = $this->configureActionButtons($action, $object);
 
         foreach ($this->getExtensions() as $extension) {
-            // TODO: remove method check in next major release
-            if (method_exists($extension, 'configureActionButtons')) {
-                $list = $extension->configureActionButtons($this, $list, $action, $object);
-            }
+            $list = $extension->configureActionButtons($this, $list, $action, $object);
         }
 
         return $list;
@@ -3125,5 +3114,44 @@ abstract class AbstractAdmin implements AdminInterface, DomainObjectInterface
         foreach ($this->getExtensions() as $extension) {
             $extension->configureRoutes($this, $this->routes);
         }
+    }
+
+    /**
+     * Setting to true will enable mosaic button for
+     * the admin screen.
+     * Setting to false will hide mosaic button for
+     * the admin screen.
+     */
+    public function setIsShowMosaicButton($isShow)
+    {
+        $var = (bool)$isShow;
+
+        if (!$var) {
+            $this->hideMosaicButton();
+            return;
+        }
+
+        $this->showMosaicButton();
+    }
+
+    protected function hideMosaicButton()
+    {
+        unset($this->listModes['mosaic']);
+    }
+
+    protected function initListModes()
+    {
+        $this->listModes = array(
+            'list' => array(
+                'class' => 'fa fa-list fa-fw',
+            )
+        );
+
+        $this->showMosaicButton();
+    }
+
+    protected function showMosaicButton()
+    {
+        $this->listModes['mosaic'] = array('class' => 'fa fa-th-large fa-fw');
     }
 }