+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.x]
+### Added
+- Add the ``configureActionButtons`` method to the AdminInterface
+- Add the ``configureActionButtons`` method to the AdminExtensionInterface
+- Add the ``configureBatchActions`` method to the AdminExtensionInterface
+
 ## [3.x]
 ### Added
 - Added AbstractAdmin, replacing Admin